savign and loading are functional with an example case in there already.
