+++ Main.py
@@ -4,12 +4,11 @@
 __author__ = 'Cowboy8625', 'Cyy', 'BJTMastermind', 'HexTree', 'Byteme8bit', 'Giioke', 'ThePinkChicken', 'Djsurry'
 
 # Import Modules
-import random
 import sys
 import time
 
-from script.Character import *
-from script import InfoDics, Items, Screen, Story
+from Mechanics.core_mechanics import *
+from script import Screen
 from Mechanics.ui_mechanics import *
 import ChangeLog
 
@@ -21,178 +20,36 @@ turn = True
 # Main() is the first function
 # Gets input from player to either Start Load Help or Exit
 def main():
-    Screen.main_menu_screen()
-    choice = input('\n\nSELECT A NUMBER:> ')
 
-    if choice == '1':
-        main_game_loop(char_creation(), opening=True)
-    elif choice == '2':  # LOADED AND SAVE
-        print('NOT AN OPTION YET')
-        time.sleep(2)
-        main()
-    elif choice == '3':  # Help
-        print('NOT AN OPTION YET')
-        time.sleep(2)
-        main()
-    elif choice == '4':
-        ChangeLog.change_log_print()
-        pause()
-        main()
-    elif choice == '5':
-        sys.exit()
-    else:
-        print("Not an Option.")
-        pause()
-        main()
+    choice = ''             # Initializes variable
 
+    while choice != '1':            # Continues loop while choice is not to start the game - will change later
+        Screen.main_menu_screen()
+        choice = input('\n\nSELECT A NUMBER:> ')
 
-# Helps is to show player how to fight use items and more
-# I think I will also make this accessable in game as well at some point
+        if choice == '1':           # Starts Game
+            main_game_loop(char_creation(), opening=True)
 
-def game_help():
-    pass
+        elif choice == '2':  # LOADED AND SAVE
+            print('NOT AN OPTION YET')
+            time.sleep(2)
+
+        elif choice == '3':  # Help
+            print('NOT AN OPTION YET')
+            time.sleep(2)
+
+        elif choice == '4':
+            ChangeLog.change_log_print()
+            pause()
+
+        elif choice == '5':
+            sys.exit()
+
+        else:
+            print("Not an Option.")
+            pause()
 
 
-# This is where the player chooses what class he will be
-def char_creation():
-    pname = None
-    pclass = None
-    pgender = None
-
-    # This function sets the name of the player
-    def player_select_name(__pname, __pclass):
-        while __pname is None:
-            message = f"A {__pclass} is a good choice I think. What shall I call you?"
-            Screen.display(message)
-            option = input("\n\nENTER YOUR NAME:> ")
-
-            if len(option) == 0:
-                print('\nYou need to enter a name to continue\n')
-            else:
-                yes_no = input("Are you Sure You Want To Keep This Name?\n"
-                                "(1): Yes\n"
-                                "(2): No\n"
-                                "Enter A Number:> ")
-
-                if yes_no == '1':
-                    __pname = option
-                    return __pname
-
-    # This set the class for the player
-    def player_select_class(__pclass):
-        player_options = ('Mage', 'Warrior', 'Archer', 'Assassin')
-        Screen.char_options()
-
-        while __pclass is None:
-            choice = input('\n\n'
-                           'SELECT A NUMBER OR TYPE HELP THEN NUMBER:> ')
-
-            if choice == '1':
-                __pclass = player_options[int(choice) - 1]
-            elif choice == '2':
-                __pclass = player_options[int(choice) - 1]
-            elif choice == '3':
-                __pclass = player_options[int(choice) - 1]
-            elif choice == '4':
-                __pclass = player_options[int(choice) - 1]
-            # lore is the story and it is found in InfoDics
-            elif choice[0:4].lower() == 'info':
-                lore = InfoDics.info_on_classes[player_options[int(choice[5]) - 1]]
-                Screen.display(lore)
-                pause()
-            else:
-                print("Not An Option.")
-        return __pclass
-
-    def player_select_gender(__pgender, __pname, __pclass):
-        while pgender is None:
-            message = f"{__pname} the {__pclass}, has a nice ring to\
-                     it I think. I know it is obvious but just let me know what gender you are."
-
-            Screen.display(message)
-            option = input('\n\n'
-                           '(1): Male\n'
-                           '(2): Female\n'
-                           'Enter A Number:> ')
-
-            if option == '1':
-                __pgender = 'male'
-
-            elif option == '2':
-                __pgender = 'female'
-
-            Story.set_gender(__pgender)
-            return __pgender
-
-    pclass = player_select_class(pclass)
-    pname = player_select_name(pname, pclass)
-    pgender = player_select_gender(pgender, pname, pclass)
-
-
-    # Mage
-    if pclass == 'Mage':
-        return Player(
-            name=pname,
-            _class=pclass,
-            max_health=80,
-            melee_attack=1,
-            magic_attack=10,
-            max_mana=50,
-            max_stamina=10,
-            defense=1,
-            luck=1,
-            gender=pgender,
-            gold=0,
-            equipped_weapon=Items.weak_staff
-        )
-    # Warrior
-    elif pclass == 'Warrior':
-        return Player(
-            name=pname,
-            _class=pclass,
-            max_health=150,
-            melee_attack=10,
-            magic_attack=0,
-            max_mana=5,
-            max_stamina=50,
-            defense=5,
-            luck=0,
-            gender=pgender,
-            gold=0,
-            equipped_weapon=Items.rusty_short_sword
-        )
-    # Archer
-    elif pclass == 'Archer':
-        return  Player(
-            name=pname,
-            _class=pclass,
-            max_health=100,
-            melee_attack=5,
-            magic_attack=0,
-            max_mana=50,
-            max_stamina=10,
-            defense=1,
-            luck=5,
-            gender=pgender,
-            gold=0,
-            equipped_weapon=Items.common_hunting_bow
-        )
-    # Assassin
-    elif pclass == 'Assassin':
-        return Player(
-            name=pname,
-            _class=pclass,
-            max_health=50,
-            melee_attack=20,
-            magic_attack=10,
-            max_mana=25,
-            max_stamina=10,
-            defense=2,
-            luck=10,
-            gender=pgender,
-            gold=100,
-            equipped_weapon=Items.rusty_dagger
-        )
 
 """
 # Makes a random mob with in the range of the players level
